refactor(orchestrator): extract settings/permissions generation to _settings.py mixin
New module:
- _settings.py: AgentSettingsMixin providing _get_role_permissions (role-based tool allow/deny) and _generate_agent_settings (per-agent settings.json generation)

orchestrator.py:
- AgentOrchestrator now inherits from AgentSettingsMixin
- Removed -355 lines (lines 343-698)
- Removed unused shutil import (moved to _settings.py)

259/259 runtime tests still pass.
