+++ roboco/runtime/orchestrator.py
@@ -15,7 +15,6 @@ The orchestrator is the BRAIN of the system:
 import asyncio
 import contextlib
 import json
-import shutil
 import tempfile
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
@@ -101,6 +100,7 @@ from roboco.runtime._helpers import (  # noqa: E402, F401
     _SlaBreach,
     get_agent_image,
 )
+from roboco.runtime._settings import AgentSettingsMixin  # noqa: E402
 
 
 class AgentReadinessError(Exception):
@@ -111,7 +111,7 @@ class AgentReadinessError(Exception):
     """
 
 
-class AgentOrchestrator:
+class AgentOrchestrator(AgentSettingsMixin):
     """
     Manages Claude Code containers for all agents.
 
@@ -339,362 +339,6 @@ class AgentOrchestrator:
     # =========================================================================
     # PER-AGENT SETTINGS GENERATION
     # =========================================================================
-
-    def _get_role_permissions(
-        self, role: str, workspace_path: str, cell_workspace_path: str
-    ) -> dict[str, list[str]]:
-        """Get role-specific allow/deny lists for Claude Code tools.
-
-        Post-gateway shape: every state-changing operation an agent can
-        perform routes through ``mcp__roboco-flow__*`` (intent verbs) or
-        ``mcp__roboco-do__*`` (content tools — commit, push, PR, journal,
-        notify, message), both granted to every role via ``base_allow``.
-        Role-specific configuration here only governs file IO (Write/Edit
-        scoping) plus a small handful of legacy native-tool denies that
-        remain meaningful for weak models. Read-only git lives in
-        ``mcp__roboco-git-readonly__*``.
-
-        Args:
-            role: Agent role (developer, qa, documenter, cell_pm, main_pm, etc.)
-            workspace_path: Path to agent's own workspace directory
-            cell_workspace_path: Path to cell's workspace root (for QA/Docs access)
-
-        Returns:
-            Dict with 'allow' and 'deny' lists for Claude Code permissions
-        """
-        # workspace_path: /data/workspaces/{project}/{team}/{agent}
-        # cell_workspace_path: /data/workspaces/{project}/{team}
-        configs: dict[str, dict[str, list[str]]] = {
-            "developer": {
-                "allow": [
-                    f"Write(/{workspace_path}/**)",
-                    f"Edit(/{workspace_path}/**)",
-                ],
-                "deny": [],
-            },
-            "qa": {
-                # QA reads code + the open PR via the gateway; never edits.
-                "allow": [],
-                "deny": [
-                    "Write(*)",
-                    "Edit(*)",
-                ],
-            },
-            "documenter": {
-                "allow": [
-                    f"Write(/{cell_workspace_path}/**)",
-                    f"Edit(/{cell_workspace_path}/**)",
-                    "Write(//app/docs/**)",
-                    "Edit(//app/docs/**)",
-                    "Write(//app/CHANGELOG.md)",
-                    "Edit(//app/CHANGELOG.md)",
-                    "Write(//app/README.md)",
-                    "Edit(//app/README.md)",
-                ],
-                "deny": [],
-            },
-            "cell_pm": {
-                # PMs coordinate; they open + merge PRs through the gateway
-                # but never author code. Edit/Write are denied so weaker
-                # models can't read the subtask title imperatively and
-                # start editing source — they have to decompose into a dev
-                # subtask. Devs are the only role that authors code.
-                "allow": [],
-                "deny": [
-                    "Bash(git commit:*)",
-                    "Bash(git push:*)",
-                    "Write(*)",
-                    "Edit(*)",
-                ],
-            },
-            "main_pm": {
-                # Same reasoning as cell_pm — Main PM sits between CEO and
-                # cell PMs; the work product is coordination + review, not
-                # commits or edits. Code work routes Main PM → Cell PM →
-                # Dev only.
-                "allow": [],
-                "deny": [
-                    "Bash(git commit:*)",
-                    "Bash(git push:*)",
-                    "Write(*)",
-                    "Edit(*)",
-                ],
-            },
-            "product_owner": {
-                "allow": [
-                    f"Write(/{workspace_path}/**)",
-                    f"Edit(/{workspace_path}/**)",
-                ],
-                "deny": [],
-            },
-            "head_marketing": {
-                "allow": [
-                    f"Write(/{workspace_path}/**)",
-                    f"Edit(/{workspace_path}/**)",
-                ],
-                "deny": [],
-            },
-            "auditor": {
-                # Auditor is read-only across the org — observes, never edits.
-                "allow": [],
-                "deny": [
-                    "Write(*)",
-                    "Edit(*)",
-                ],
-            },
-        }
-
-        if role not in configs:
-            logger.warning(
-                "No Claude Code permissions configured for role; "
-                "agent will be limited to base_allow/base_deny.",
-                role=role,
-            )
-        return configs.get(role, {"allow": [], "deny": []})
-
-    def _generate_agent_settings(
-        self,
-        agent_id: str,
-        role: str,
-        workspace_path: str,
-        cell_workspace_path: str,
-    ) -> Path:
-        """Generate per-agent Claude Code settings file with role-specific permissions.
-
-        This replaces the shared settings approach. Each agent gets their own
-        settings.json with:
-        - Base MCP tools allowed for all agents
-        - Role-specific tool permissions
-        - Explicit deny list blocking native git/file operations
-
-        Args:
-            agent_id: Agent identifier (e.g., "be-dev-1")
-            role: Agent role (e.g., "developer")
-            workspace_path: Path to agent's own workspace directory
-            cell_workspace_path: Path to cell's workspace root (for QA/Docs)
-
-        Returns:
-            Path to the generated settings file
-        """
-        # Base MCP tools for all agents. Post-gateway every role gets the
-        # full intent-verb + content-tool surface; the orchestrator-side
-        # API rejects verbs/tools the agent's role isn't authorized for,
-        # so granting `*` here is safe.
-        base_allow = [
-            "mcp__roboco-flow__*",
-            "mcp__roboco-do__*",
-            "mcp__roboco-optimal__*",
-            "mcp__roboco-git-readonly__*",
-            "Read(*)",  # All agents can read any file
-        ]
-
-        # Base denials for all agents - block native tools + sensitive reads.
-        # The Read/Bash denies below are critical: without them an agent can
-        # read `.git/config` (which, pre-fix, had the PAT embedded in the
-        # remote URL) or `~/.gitconfig` and exfiltrate project secrets.
-        # We also block direct curl/wget to github.com — any git-remote op
-        # must go through the orchestrator's git service, which injects the
-        # token via bearer header at subprocess time rather than exposing it.
-        base_deny = [
-            # Block ALL native git commands - must use roboco_git_* tools
-            "Bash(git:*)",
-            # NOTE: Write/Edit are intentionally NOT globally denied here.
-            # Claude Code evaluates rules deny -> ask -> allow and the first
-            # match wins, so a deny ALWAYS beats a more-specific allow (the
-            # glob syntax has no negation). A global Write(*)/Edit(*) here
-            # therefore unconditionally shadowed the per-role,
-            # workspace-scoped Write/Edit allows below — every agent (devs
-            # included) was unable to edit ANY file and fell back to
-            # destructive bash redirection (clobbering real files). Roles
-            # that must NOT write (qa, cell_pm, main_pm, auditor) carry
-            # their own Write(*)/Edit(*) deny in _get_role_permissions.
-            # Block reads of credential stores, anywhere on the FS
-            "Read(**/.git/config)",
-            "Read(**/.gitconfig)",
-            "Read(/etc/gitconfig)",
-            "Read(~/.netrc)",
-            "Read(**/.git-credentials)",
-            # Block direct GitHub API/wire access — agents must use
-            # roboco_git_* MCP tools so secrets + traceability stay on the
-            # orchestrator side.
-            "Bash(curl:*github.com*)",
-            "Bash(curl:*api.github.com*)",
-            "Bash(wget:*github.com*)",
-            "Bash(wget:*api.github.com*)",
-            # Same idea for cat-ing credential files in a subshell
-            "Bash(cat:*.git/config*)",
-            "Bash(cat:*.gitconfig*)",
-            "Bash(cat:*.git-credentials*)",
-            # Block reading env vars that might leak secrets
-            "Bash(env:*)",
-            "Bash(printenv:*)",
-        ]
-
-        # Get role-specific permissions
-        role_config = self._get_role_permissions(
-            role, workspace_path, cell_workspace_path
-        )
-
-        # Combine base + role-specific.
-        # defaultMode=bypassPermissions lets unlisted operations proceed
-        # without an interactive prompt (which would hang a non-TTY agent
-        # container). Explicit deny rules still apply.
-        settings: dict[str, Any] = {
-            "permissions": {
-                "defaultMode": "bypassPermissions",
-                "allow": base_allow + role_config["allow"],
-                "deny": base_deny + role_config["deny"],
-            },
-            "hooks": {
-                # Start SDK server on session start (for A2A communication)
-                "SessionStart": [
-                    {
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/sdk-startup-hook.sh",
-                            }
-                        ]
-                    }
-                ],
-                # Guard Bash: block shell-level git/curl/wget/env patterns
-                # that the matcher-based `permissions.deny` can't catch
-                # (e.g. `cd X && git fetch`). Redirects agents to the MCP
-                # equivalents instead of bloating prompts with rules.
-                "PreToolUse": [
-                    {
-                        "matcher": "Bash",
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/bash-guard-hook.sh",
-                            }
-                        ],
-                    },
-                ],
-                "PostToolUse": [
-                    # Check for incoming A2A messages after each tool use
-                    {
-                        "matcher": "*",
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/a2a-check-hook.sh",
-                            }
-                        ],
-                    },
-                    # Per-session budget counter + loop detector. Shared SDK
-                    # state lets this hook emit [Budget]/[Loop]/[Halt]
-                    # reminders that the orchestrator's kill-switch corroborates.
-                    {
-                        "matcher": "*",
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/post-tool-budget-hook.sh",
-                            }
-                        ],
-                    },
-                    # Sync token usage from the transcript so /usage/status
-                    # (and the cost dashboard) reflect real spend. Idempotent
-                    # absolute set — running it per tool keeps mid-run
-                    # snapshots and reaped-agent sessions accurate.
-                    {
-                        "matcher": "*",
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/usage-report-hook.sh",
-                            }
-                        ],
-                    },
-                ],
-                # Stop guard: refuse silent exits unless a terminal tool was
-                # just called (idle/substitute/escalate/pause/...). Second
-                # attempt auto-substitutes via SDK so the task doesn't rot.
-                "Stop": [
-                    {
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/stop-hook.sh",
-                            },
-                            # Final token-usage sync at turn end — guarantees
-                            # the session total is captured before the agent
-                            # idles and the orchestrator finalizes the row.
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/usage-report-hook.sh",
-                            },
-                        ]
-                    }
-                ],
-                # Prompt-injection guard — rejects turns that look like
-                # another agent's content trying to override our rules.
-                "UserPromptSubmit": [
-                    {
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/user-prompt-hook.sh",
-                            }
-                        ]
-                    }
-                ],
-                # Snapshot budget / terminal state before compact so the
-                # next session resumes with continuity.
-                "PreCompact": [
-                    {
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/pre-compact-hook.sh",
-                            }
-                        ]
-                    }
-                ],
-                # Post-mortem: write a reflect-journal entry summarising the
-                # session (tools called, halt/loop triggered, last tool).
-                "SessionEnd": [
-                    {
-                        "hooks": [
-                            {
-                                "type": "command",
-                                "command": "/app/scripts/session-end-hook.sh",
-                            }
-                        ]
-                    }
-                ],
-            },
-        }
-
-        # Write to per-agent settings file
-        # When running in container: write to /app/agent-settings (mounted to host)
-        # When running on host: use temp directory
-        if DATA_HOST_PATH:
-            settings_dir = Path("/app/agent-settings")
-        else:
-            settings_dir = Path(tempfile.gettempdir()) / "roboco-agent-settings"
-
-        settings_dir.mkdir(parents=True, exist_ok=True)
-        settings_path = settings_dir / f"{agent_id}-settings.json"
-
-        # Handle case where Docker auto-created a directory instead of a file
-        if settings_path.is_dir():
-            shutil.rmtree(settings_path)
-
-        settings_path.write_text(json.dumps(settings, indent=2))
-
-        logger.debug(
-            "Generated per-agent settings",
-            agent_id=agent_id,
-            role=role,
-            settings_path=str(settings_path),
-            allow_count=len(settings["permissions"]["allow"]),
-            deny_count=len(settings["permissions"]["deny"]),
-        )
-
-        return settings_path
 
     # =========================================================================
     # AGENT SPAWNING
